requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-amazon==3.0.0
psycopg2-binary
boto3==1.24.9
confluent-kafka==2.2.0
kafka-python
python-dotenv

